requirements.txt
streamlit==0.82.0
opencv_python==4.4.0.42
numpy==1.18.5
scipy==1.4.1
onnxruntime_gpu==1.9.0
imutils==0.5.4
onnxruntime==1.9.0
pyrebase==3.0.27
PyYAML==5.4.1
